+++ app/src/main/java/com/technzone/miniborsa/data/models/business/businessrequest/BusinessRequest.kt
@@ -37,7 +37,7 @@ data class BusinessRequest(
     val isRelocated: Boolean? = null,
 
     @field:SerializedName("askingPriceBoth")
-    val askingPriceBoth: Int? = null,
+    val askingPriceBoth: Double? = null,
 
     @field:SerializedName("investmentPercentage")
     val investmentPercentage: Int? = null,
@@ -52,7 +52,7 @@ data class BusinessRequest(
     val askingPriceNABoth: Boolean? = null,
 
     @field:SerializedName("annualTurnover")
-    val annualTurnover: Int? = null,
+    val annualTurnover: Double? = null,
 
     @field:SerializedName("listLocation")
     val listLocation: Boolean? = null,
@@ -76,7 +76,7 @@ data class BusinessRequest(
     val isConfidential: Boolean? = null,
 
     @field:SerializedName("askingPrice")
-    val askingPrice: Int? = null,
+    val askingPrice: Double? = null,
 
     @field:SerializedName("askingPriceNA")
     val askingPriceNA: Boolean? = null,
@@ -88,7 +88,7 @@ data class BusinessRequest(
     val countries: List<Int?>? = null,
 
     @field:SerializedName("annualNetProfit")
-    val annualNetProfit: Int? = null,
+    val annualNetProfit: Double? = null,
 
     @field:SerializedName("annualTurnoverNA")
     val annualTurnoverNA: Boolean? = null,
